// src/lib/pb/db-types.ts
type AuthCollectionRecord = {
  id: string;
  created: string;
  updated: string;
  username: string;
  email: string;
  emailVisibility: boolean;
  verified: boolean;
};
// ...
export type UtilityTenantsResponse = {
  phone?: string;
  avatar?:string; 
} & AuthCollectionRecord;

export type UtilityTenantsCreate = {
  phone?: string;
  avatar?: string|File;
};